--- a/src/nanovg_vk.h
+++ b/src/nanovg_vk.h
@@ -17,7 +17,9 @@ typedef struct VKNVGCreateInfo {
   VkPhysicalDevice gpu;
   VkDevice device;
   VkRenderPass renderpass;
-  VkCommandBuffer cmdBuffer;
+  VkCommandBuffer *cmdBuffer;
+  uint32_t swapChainImageCount;
+  uint32_t *currentBuffer;
 
   const VkAllocationCallbacks *allocator; //Allocator for vulkan. can be null
 } VKNVGCreateInfo;
@@ -1056,7 +1058,8 @@ static void vknvg_fill(VKNVGcontext *vk, VKNVGcall *call) {
   int i, npaths = call->pathCount;
 
   VkDevice device = vk->createInfo.device;
-  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer;
+  uint32_t currentBuffer = *vk->createInfo.currentBuffer;
+  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer[currentBuffer];
 
 #ifdef __cplusplus
   VKNVGCreatePipelineKey pipelinekey = {};
@@ -1126,7 +1129,8 @@ static void vknvg_convexFill(VKNVGcontext *vk, VKNVGcall *call) {
   int npaths = call->pathCount;
 
   VkDevice device = vk->createInfo.device;
-  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer;
+  uint32_t currentBuffer = *vk->createInfo.currentBuffer;
+  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer[currentBuffer];
   
 #ifdef __cplusplus
   VKNVGCreatePipelineKey pipelinekey = {};
@@ -1172,9 +1176,10 @@ static void vknvg_convexFill(VKNVGcontext *vk, VKNVGcall *call) {
 
 static void vknvg_stroke(VKNVGcontext *vk, VKNVGcall *call) {
   VkDevice device = vk->createInfo.device;
-  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer;
+  uint32_t currentBuffer = *vk->createInfo.currentBuffer;
+  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer[currentBuffer];
 
-  VKNVGpath *paths = &vk->paths[call->pathOffset];
+    VKNVGpath *paths = &vk->paths[call->pathOffset];
   int npaths = call->pathCount;
 
   if (vk->flags & NVG_STENCIL_STROKES) {
@@ -1270,7 +1275,8 @@ static void vknvg_triangles(VKNVGcontext *vk, VKNVGcall *call) {
     return;
   }
   VkDevice device = vk->createInfo.device;
-  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer;
+  uint32_t currentBuffer = *vk->createInfo.currentBuffer;
+  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer[currentBuffer];
 
 #ifdef __cplusplus
   VKNVGCreatePipelineKey pipelinekey = {};
@@ -1460,8 +1466,9 @@ static int vknvg_renderCreateTexture(void *uptr, int type, int w, int h, int ima
     vknvg_UpdateTexture(device, tex, 0, 0, w, h, generated_texture);
     free(generated_texture);
   }
-  
-  vknvg_InitTexture(vk->createInfo.cmdBuffer, vk->queue, tex);
+
+  uint32_t currentBuffer = *vk->createInfo.currentBuffer;
+  vknvg_InitTexture(vk->createInfo.cmdBuffer[currentBuffer], vk->queue, tex);
 
   return vknvg_textureId(vk, tex);
 }
@@ -1507,8 +1514,8 @@ static void vknvg_renderCancel(void *uptr) {
 static void vknvg_renderFlush(void *uptr) {
   VKNVGcontext *vk = (VKNVGcontext *)uptr;
   VkDevice device = vk->createInfo.device;
-  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer;
-  VkRenderPass renderpass = vk->createInfo.renderpass;
+  uint32_t currentBuffer = *vk->createInfo.currentBuffer;
+  VkCommandBuffer cmdBuffer = vk->createInfo.cmdBuffer[currentBuffer];
   VkPhysicalDeviceMemoryProperties memoryProperties = vk->memoryProperties;
   const VkAllocationCallbacks *allocator = vk->createInfo.allocator;