Hi Everyone.

This PR is a followup to a comment I left a few months ago, visible here https://github.com/memononen/nanovg/issues/614#issuecomment-1426870962

I made some optimizations that have doubled the frame-rate of the `example_vulkan` demo on my machine. These changes do not break compatibility with the older implementations either, specifically with `example_vulkan_min_no_glfw`.

The changes I made are listed below:

- Support for multiple command buffers, specifically one per swap-chain image / buffer
- Caching of descriptor sets, removing the need to create new ones per draw func or to call `vkResetDescriptorPool` per frame.
- Under `example_vulkan`, using `vkWaitForFences` to control rendering as opposed to `vkQueueWaitIdle` (seems to be the biggest perf booster). This change also has implied "multiple frames in flight" as dictated by the swap-chain image count.
- Using persisted mapped buffers via `VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT` instead of `VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT `, allows to skip calls to `vkMap/UnmapMemory` per frame

The best way to observe the differences in performance is to run `example_vulkan` and `example_vulkan_min_no_glfw` separately. On my system ( RX 6800, Ryzen R7 7700 ) the `example_vulkan` demo has an average frame time of 0.80 ms (1250 fps) as opposed to the `example_vulkan_min_no_glfw` demo with 1.98 ms (505 fps). For reference `example-gl3` runs at 0.55 ms (1818 fps), but this is a big improvement as `example_vulkan` is now just 31% slower as opposed to 72% slower for `example_vulkan_min_no_glfw`.

The main difference in terms of setup is:

- Treating `cmd_buffer` as an array
- Setting `swapchainImageCount` and `* currentBuffer` within the `VKNVGCreateInfo` init. 

Open for comments and feedback, particular around tests on Nvidia, Intel and Apple Silicon hardware.

![Screenshot 2023-06-12 115445](https://github.com/danilw/nanovg_vulkan/assets/16505156/e2f946a3-c897-40ef-9901-8a1eb202d9b0)

![Screenshot 2023-06-12 115528](https://github.com/danilw/nanovg_vulkan/assets/16505156/164b8d91-875e-47a8-82e4-ea2819b85699)