@danilw 2023-06-23
> These changes do not break compatibility with the older implementations either, specifically with example_vulkan_min_no_glfw.

on Linux I can not build your repo after cloning, there few mistake in code `example/example_vulkan_min_no_glfw.c`, I think you change and not tested Linux version

can you change your `example/example_vulkan_min_no_glfw.c` to this fixed, in attachment
there 3 changes at 450-480 lines arroung `VK_USE_PLATFORM_XCB_KHR`
[example_vulkan_min_no_glfw.zip](https://github.com/danilw/nanovg_vulkan/files/11848066/example_vulkan_min_no_glfw.zip)