@danilw 2023-06-23
### About performance:

**Nvidia** - **absolutely no change in performance**, your updated repo - it is exact same to this my original without your changes, I opened multiple window, make them fullscreen, pressed Space to test - everything is exact same. 
And Nvidia performance with this is "same bad" as before about 50% slower than OpenGL. Nvidia drop below 60FSP on fullscreen with Space-clicked.
*But maybe this is because I use Wayland with Nvidia being second GPU and Nvidia performance just "downgraded" because Nvidia driver being bad...*

**AMD** - **huge improvement** on this repository-examples. from ~300fps original to 1200fps your improved.
it is even 2x more fps than "multiple frames in flight" example.

Now il check "multiple frames in flight"... I thought it will be faster...