@SubiyaCryolite 2023-06-24
> If I understand correctly - you integrated "multiple frames in flight" to `nanovg_vk.h` but you did it more correctly than I did in my example, you did it with cache and and memory optimizations. Looks good. You also keep compatibility with "minimal example that use single frame".
> 
> I need to remake my external examples with integration, il do it latter.
> 
> **Il wait for you reply.**
> 
> **Confirm if you do this fixes or I will after accepting your pull request:**
> 
>     1. Linux fix in [Optimizations > Cached Descriptor Sets, Implied Multiple Frames in Flight, Fencing for faster perf 2X #7 (comment)](https://github.com/danilw/nanovg_vulkan/pull/7#issuecomment-1604170006)
> 
>     2. I also saw g++ complaining about this:
>        https://github.com/SubiyaCryolite/nanovg_vulkan/blob/feature/use-persistent-mapped-memory/src/nanovg_vk.h#L1504 arguments of `vknvg_UpdateBuffer` :
>        `VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT` 
>        when:
>        `static void vknvg_UpdateBuffer(...` **`VkMemoryPropertyFlagBits memory_type`** `...)`
> 
> 
> **I think it may work incorrectly** because expected `VkMemoryPropertyFlagBits` and VkMemoryPropertyFlagBits is `typedef enum VkMemoryPropertyFlagBits` https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkMemoryPropertyFlagBits.html
> 
> And your value result of `VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT` may not go correctly.
> 
> **I think this need to be fixed**, by replacing `VkMemoryPropertyFlagBits memory_type` with `uint32_t` type or `VkFlags` type in chain down to `vknvg_memory_type_from_properties`.
> 
> I have not tested it on Windows, I assume it works. **I tested with validation layers on Nvidia and AMD - no errors**. (there is 1 error when you close example_vulkan something is not cleaned - but it is same as in my original, I just keep it as it is this is not important)
> 
> **I also saw bug with resize - crash on resize very common**, on AMD GPU only, but my minimal "vulkan shader app" that is source of "no_glfw" example does not crash here, so this need investigation... There no validation errors on crash it just segfault, crash message is this:
> 
> ```
> example-vk: /home/danil/2021_vulkan_projects/not_clean/nanovg_vulkan/example/example_vulkan.c:66: prepareFrame: Assertion `res == VK_SUCCESS' failed.
> ```
> 
> or
> 
> ```
> example-vk: /home/danil/2021_vulkan_projects/not_clean/nanovg_vulkan/example/example_vulkan.c:171: submitFrame: Assertion `res == VK_SUCCESS' failed.
> ```
> 
> **But same bug happening in this my original version, not your bug.**

Thanks for the feedback. I'll be sure to look into these issues in the upcoming days.

Unfortunately I only have access to an AMD GPU (Windows) and my M1 Macbook Pro. So I may not be able to test on Nvidia anytime soon.

Thanks again.