@danilw 2023-06-24
> Unfortunately I only have access to an AMD GPU (Windows) and my M1 Macbook Pro. So I may not be able to test on Nvidia anytime soon.

I mean - its fine, il test Linux and Nvidia